+++ packages/vidstack/src/components/ui/thumbnails/thumbnail.ts
@@ -1,23 +1,10 @@
-import { Component, effect, peek, State } from 'maverick.js';
-import { animationFrameThrottle, isNull, listenEvent } from 'maverick.js/std';
-import type { VTTCue } from 'media-captions';
+import { Component, effect, State } from 'maverick.js';
+import { isNull, listenEvent } from 'maverick.js/std';
 
 import { useMediaContext, type MediaContext } from '../../../core/api/media-context';
 import type { MediaCrossOrigin } from '../../../core/api/types';
-import { findActiveCue } from '../../../core/tracks/text/utils';
 import { $ariaBool } from '../../../utils/aria';
-import { ThumbnailsLoader } from './thumbnail-loader';
-
-export interface ThumbnailState {
-  src: string;
-  img: HTMLImageElement | null | undefined;
-  crossOrigin: MediaCrossOrigin | null;
-  coords: ThumbnailCoords | null;
-  activeCue: VTTCue | null;
-  loading: boolean;
-  error: ErrorEvent | null;
-  hidden: boolean;
-}
+import { ThumbnailsLoader, type ThumbnailImage, type ThumbnailSrc } from './thumbnail-loader';
 
 /**
  * Used to load and display a preview thumbnail at the given `time`.
@@ -29,7 +16,7 @@ export interface ThumbnailState {
  */
 export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
   static props: ThumbnailProps = {
-    src: '',
+    src: null,
     time: 0,
     crossOrigin: null,
   };
@@ -37,8 +24,8 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
   static state = new State<ThumbnailState>({
     src: '',
     img: null,
-    coords: null,
-    activeCue: null,
+    thumbnails: [],
+    activeThumbnail: null,
     crossOrigin: null,
     loading: false,
     error: null,
@@ -46,13 +33,13 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
   });
 
   protected _media!: MediaContext;
-  protected _thumbnails!: ThumbnailsLoader;
+  protected _loader!: ThumbnailsLoader;
 
   private _styleResets: (() => void)[] = [];
 
   protected override onSetup(): void {
     this._media = useMediaContext();
-    this._thumbnails = ThumbnailsLoader.create(this.$props.src, this.$state.crossOrigin);
+    this._loader = ThumbnailsLoader.create(this.$props.src, this.$state.crossOrigin);
 
     this._watchCrossOrigin();
 
@@ -69,8 +56,8 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
     effect(this._watchHidden.bind(this));
     effect(this._watchCrossOrigin.bind(this));
     effect(this._onLoadStart.bind(this));
-    effect(this._onFindActiveCue.bind(this));
-    effect(this._onResolveThumbnail.bind(this));
+    effect(this._onFindActiveThumbnail.bind(this));
+    effect(this._resize.bind(this));
   }
 
   private _watchImg() {
@@ -90,16 +77,23 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
 
   private _onLoadStart() {
     const { src, loading, error } = this.$state;
-    src();
-    loading.set(true);
-    error.set(null);
+
+    if (src()) {
+      loading.set(true);
+      error.set(null);
+    }
+
+    return () => {
+      this._resetStyles();
+      loading.set(false);
+      error.set(null);
+    };
   }
 
   private _onLoaded() {
     const { loading, error } = this.$state;
     loading.set(false);
     error.set(null);
-    this._requestResize();
   }
 
   private _onError(event: ErrorEvent) {
@@ -121,93 +115,69 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
   private _watchHidden() {
     const { hidden } = this.$state,
       { duration } = this._media.$state,
-      cues = this._thumbnails.$cues();
-    hidden.set(this._hasError() || !Number.isFinite(duration()) || cues.length === 0);
+      images = this._loader.$images();
+    hidden.set(this._hasError() || !Number.isFinite(duration()) || images.length === 0);
   }
 
   protected _getTime() {
     return this.$props.time();
   }
 
-  private _onFindActiveCue() {
-    const time = this._getTime(),
-      { activeCue } = this.$state,
-      { duration } = this._media.$state,
-      cues = this._thumbnails.$cues();
+  private _onFindActiveThumbnail() {
+    let time = this._getTime(),
+      { src, activeThumbnail } = this.$state,
+      images = this._loader.$images(),
+      activeIndex = -1,
+      activeImage: ThumbnailImage | null = null;
 
-    if (!cues || !Number.isFinite(duration())) {
-      activeCue.set(null);
-      return;
+    for (let i = images.length - 1; i >= 0; i--) {
+      const image = images[i];
+      if (time >= image.startTime && (!image.endTime || time < image.endTime)) {
+        activeIndex = i;
+        break;
+      }
     }
 
-    activeCue.set(findActiveCue(cues, time));
-  }
-
-  private _onResolveThumbnail() {
-    let { activeCue } = this.$state,
-      cue = activeCue(),
-      baseURL = peek(this.$props.src);
-
-    if (!/^https?:/.test(baseURL)) {
-      baseURL = location.href;
+    if (images[activeIndex]) {
+      activeImage = images[activeIndex];
     }
 
-    if (!baseURL || !cue) {
-      this.$state.src.set('');
-      this._resetStyles();
-      return;
-    }
-
-    const [src, coords = ''] = (cue.text || '').split('#');
-    this.$state.coords.set(this._resolveThumbnailCoords(coords));
-
-    this.$state.src.set(this._resolveThumbnailSrc(src, baseURL));
-    this._requestResize();
+    activeThumbnail.set(activeImage);
+    src.set(activeImage?.url.href || '');
   }
-
-  private _resolveThumbnailSrc(src: string, baseURL: string) {
-    return /^https?:/.test(src) ? src : new URL(src, baseURL).href;
-  }
-
-  private _resolveThumbnailCoords(coords: string) {
-    const [props, values] = coords.split('='),
-      resolvedCoords = {},
-      coordValues = values?.split(',');
-
-    if (!props || !values) return null;
-
-    for (let i = 0; i < props.length; i++) resolvedCoords[props[i]] = +coordValues[i];
-
-    return resolvedCoords as ThumbnailCoords;
-  }
-
-  private _requestResize = animationFrameThrottle(this._resize.bind(this));
 
   private _resize() {
     if (!this.scope) return;
 
-    const img = this.$state.img(),
-      coords = this.$state.coords();
+    const rootEl = this.el,
+      imgEl = this.$state.img(),
+      loading = this.$state.loading(),
+      thumbnail = this.$state.activeThumbnail();
 
-    if (!img || !this.el) return;
+    if (!imgEl || !thumbnail || !rootEl || loading) return;
 
-    const w = coords?.w ?? img.naturalWidth,
-      h = coords?.h ?? img.naturalHeight,
+    const width = thumbnail.width ?? imgEl.naturalWidth,
+      height = thumbnail?.height ?? imgEl.naturalHeight,
       { maxWidth, maxHeight, minWidth, minHeight } = getComputedStyle(this.el),
-      minRatio = Math.max(parseInt(minWidth) / w, parseInt(minHeight) / h),
-      maxRatio = Math.min(parseInt(maxWidth) / w, parseInt(maxHeight) / h),
+      minRatio = Math.max(parseInt(minWidth) / width, parseInt(minHeight) / height),
+      maxRatio = Math.min(
+        Math.max(parseInt(minWidth), parseInt(maxWidth)) / width,
+        Math.max(parseInt(minHeight), parseInt(maxHeight)) / height,
+      ),
       scale = maxRatio < 1 ? maxRatio : minRatio > 1 ? minRatio : 1;
 
-    this._style(this.el!, '--thumbnail-width', `${w * scale}px`);
-    this._style(this.el!, '--thumbnail-height', `${h * scale}px`);
-    this._style(img, 'width', `${img.naturalWidth * scale}px`);
-    this._style(img, 'height', `${img.naturalHeight * scale}px`);
+    this._style(rootEl, '--thumbnail-width', `${width * scale}px`);
+    this._style(rootEl, '--thumbnail-height', `${height * scale}px`);
+    this._style(imgEl, 'width', `${imgEl.naturalWidth * scale}px`);
+    this._style(imgEl, 'height', `${imgEl.naturalHeight * scale}px`);
     this._style(
-      img,
+      imgEl,
       'transform',
-      coords ? `translate(-${coords.x * scale}px, -${coords.y * scale}px)` : '',
+      thumbnail.coords
+        ? `translate(-${thumbnail.coords.x * scale}px, -${thumbnail.coords.y * scale}px)`
+        : '',
     );
-    this._style(img, 'max-width', 'none');
+    this._style(imgEl, 'max-width', 'none');
   }
 
   private _style(el: HTMLElement, name: string, value: string) {
@@ -223,10 +193,11 @@ export class Thumbnail extends Component<ThumbnailProps, ThumbnailState> {
 
 export interface ThumbnailProps {
   /**
-   * The absolute or relative URL to a [WebVTT](https://developer.mozilla.org/en-US/docs/Web/API/WebVTT_API)
-   * file resource.
+   * The thumbnails resource.
+   *
+   * @see {@link https://www.vidstack.io/docs/player/core-concepts/loading#thumbnails}
    */
-  src: string;
+  src: ThumbnailSrc;
   /**
    * Finds, loads, and displays the first active thumbnail cue that's start/end times are in range.
    */
@@ -240,9 +211,13 @@ export interface ThumbnailProps {
   crossOrigin: true | MediaCrossOrigin | null;
 }
 
-export interface ThumbnailCoords {
-  w: number;
-  h: number;
-  x: number;
-  y: number;
+export interface ThumbnailState {
+  src: string;
+  img: HTMLImageElement | null | undefined;
+  crossOrigin: MediaCrossOrigin | null;
+  thumbnails: ThumbnailImage[];
+  activeThumbnail: ThumbnailImage | null;
+  loading: boolean;
+  error: ErrorEvent | null;
+  hidden: boolean;
 }